+++ tests/Imkery.Application.UnitTests/Common/Behaviors/ValidationBehaviorTests.cs
@@ -2,44 +2,44 @@ using ErrorOr;
 using FluentAssertions;
 using FluentValidation;
 using FluentValidation.Results;
-using Imkery.Application.Apiaries.Commands.CreateApiary;
 using Imkery.Application.Common.Behaviors;
+using Imkery.Application.Apiaries.Commands.CreateApiary;
 using Imkery.Domain.Apiaries;
-using Imkery.TestShared;
-using Imkery.TestShared.Apiaries.Commands;
 using MediatR;
 using NSubstitute;
-using Xunit;
+using Imkery.TestShared.Apiaries;
 
 namespace Imkery.Application.UnitTests;
 
 public class ValidationBehaviorTests
 {
+    private readonly RequestHandlerDelegate<ErrorOr<Apiary>> _nextBehavior;
+    private readonly IValidator<CreateApiaryCommand> _validator;
+    private readonly ValidationBehavior<CreateApiaryCommand, ErrorOr<Apiary>> _sut;
+
+    public ValidationBehaviorTests()
+    {
+        _nextBehavior = Substitute.For<RequestHandlerDelegate<ErrorOr<Apiary>>>();
+        _validator = Substitute.For<IValidator<CreateApiaryCommand>>();
+
+        _sut = new ValidationBehavior<CreateApiaryCommand, ErrorOr<Apiary>>(_validator);
+    }
+
     [Fact]
     public async Task InvokeBehavior_ShouldInvokeNextBehavior_WhenValidatorResultIsValid()
     {
         // Arrange
-
-        // Create a request
         var command = ApiaryCommandsFactory.CreateApiaryCommand();
-        
-        // Create a next behavior
-        var apiary = new ApiaryBuilder().Build();
-        var nextBehavior = Substitute.For<RequestHandlerDelegate<ErrorOr<Apiary>>>();
-        nextBehavior.Invoke().Returns(apiary);
-        
-        // Create validation behavior 
-        var validator = Substitute.For<IValidator<CreateApiaryCommand>>();
-        validator
+        var apiary = new ApiaryBuilder().Build();        
+
+        _validator
             .ValidateAsync(command, Arg.Any<CancellationToken>())
             .Returns(new ValidationResult());
 
-        var validationBehavior = new ValidationBehavior<CreateApiaryCommand, ErrorOr<Apiary>>(validator);
-
-        // validator.ValidateAsync(command, Args.Any(CancellationToken));
+        _nextBehavior.Invoke().Returns(apiary);
 
         // Act 
-        var result = await validationBehavior.Handle(command, nextBehavior, Arg.Any<CancellationToken>());
+        var result = await _sut.Handle(command, _nextBehavior, Arg.Any<CancellationToken>());
 
         // Assert
         result.IsError.Should().BeFalse();
