+++ tests/Imkery.Application.UnitTests/Imkery.Application.UnitTests.csproj
@@ -18,6 +18,10 @@
   </ItemGroup>
 
   <ItemGroup>
+    <Using Include="Xunit" />
+  </ItemGroup>
+
+  <ItemGroup>
     <ProjectReference Include="..\..\src\Imkery.Application\Imkery.Application.csproj" />
     <ProjectReference Include="..\Imkery.TestShared\Imkery.TestShared.csproj" />
   </ItemGroup>
