+++ tests/Imkery.TestShared/Apiaries/ApiaryCommandsFactory.cs
@@ -1,11 +1,11 @@
 using Imkery.Application.Apiaries.Commands.CreateApiary;
 
-namespace Imkery.TestShared.Apiaries.Commands;
+namespace Imkery.TestShared.Apiaries;
 
 public static class ApiaryCommandsFactory
 {
     public static CreateApiaryCommand CreateApiaryCommand(
-        string name = "",
+        string name = "test_apiary",
         decimal latitude = 1,
         decimal longitude = 1
     )
